requirements.txt
anthropic>=0.40.0
watchdog>=4.0.0
python-docx>=1.1.0
pdfplumber>=0.11.0
PyPDF2>=3.0.0
playwright>=1.45.0
gspread>=6.0.0
google-auth>=2.30.0
google-auth-oauthlib>=1.2.0
requests>=2.32.0
python-dotenv>=1.0.0
schedule>=1.2.0
Pillow>=10.0.0
reportlab>=4.0.0
flask>=3.0.0
flask-cors>=4.0.0
gunicorn>=22.0.0
